split NextIndexAssigner
