+++ src/factor_instances_provider/next_index_assigner/mod.rs
@@ -1,3 +1,11 @@
 mod next_derivation_entity_index_assigner;
+mod next_derivation_entity_index_profile_analyzing_assigner;
+mod next_derivation_entity_index_with_local_offsets;
+mod next_derivation_entity_index_with_local_offsets_for_factor_source;
+mod offset_from_cache;
 
 pub use next_derivation_entity_index_assigner::*;
+pub use next_derivation_entity_index_profile_analyzing_assigner::*;
+pub use next_derivation_entity_index_with_local_offsets::*;
+pub use next_derivation_entity_index_with_local_offsets_for_factor_source::*;
+pub use offset_from_cache::*;
