+++ src/factor_instances_provider/next_index_assigner/next_derivation_entity_index_assigner.rs
@@ -1,264 +1,10 @@
-use std::ops::AddAssign;
-
-use crate::{factor_instances_provider::agnostic_paths, prelude::*};
-
-pub struct NextDerivationEntityIndexProfileAnalyzingAssigner {
-    network_id: NetworkID,
-
-    /// might be empty
-    unsecurified_accounts_on_network: IndexSet<UnsecurifiedEntity>,
-
-    /// might be empty
-    securified_accounts_on_network: IndexSet<SecurifiedAccount>,
-
-    /// might be empty
-    unsecurified_identities_on_network: IndexSet<UnsecurifiedEntity>,
-
-    /// might be empty
-    securified_identities_on_network: IndexSet<SecurifiedPersona>,
-}
-
-impl NextDerivationEntityIndexProfileAnalyzingAssigner {
-    pub fn new(network_id: NetworkID, profile: Option<Profile>) -> Self {
-        let unsecurified_accounts_on_network = profile
-            .as_ref()
-            .map(|p| p.unsecurified_accounts_on_network(network_id))
-            .unwrap_or_default();
-
-        let securified_accounts_on_network = profile
-            .as_ref()
-            .map(|p| p.securified_accounts_on_network(network_id))
-            .unwrap_or_default();
-
-        let unsecurified_identities_on_network = profile
-            .as_ref()
-            .map(|p| p.unsecurified_identities_on_network(network_id))
-            .unwrap_or_default();
-
-        let securified_identities_on_network = profile
-            .as_ref()
-            .map(|p| p.securified_identities_on_network(network_id))
-            .unwrap_or_default();
-
-        Self {
-            network_id,
-            unsecurified_accounts_on_network,
-            securified_accounts_on_network,
-            unsecurified_identities_on_network,
-            securified_identities_on_network,
-        }
-    }
-
-    fn max_account_veci(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-    ) -> Option<HDPathComponent> {
-        self.unsecurified_accounts_on_network
-            .clone()
-            .into_iter()
-            .map(|x: UnsecurifiedEntity| x.veci().factor_instance())
-            .filter(|f| f.factor_source_id == factor_source_id)
-            .map(|f| f.derivation_path())
-            .map(|p| {
-                AssertMatches {
-                    network_id: self.network_id,
-                    key_kind: CAP26KeyKind::TransactionSigning,
-                    entity_kind: CAP26EntityKind::Account,
-                    key_space: KeySpace::Unsecurified,
-                }
-                .matches(&p)
-            })
-            .map(|fi| fi.index)
-            .max()
-    }
-
-    pub fn max_identity_veci(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-    ) -> Option<HDPathComponent> {
-        self.unsecurified_identities_on_network
-            .clone()
-            .into_iter()
-            .map(|x: UnsecurifiedEntity| x.veci().factor_instance())
-            .filter(|f| f.factor_source_id == factor_source_id)
-            .map(|f| f.derivation_path())
-            .map(|p| {
-                AssertMatches {
-                    network_id: self.network_id,
-                    key_kind: CAP26KeyKind::TransactionSigning,
-                    entity_kind: CAP26EntityKind::Identity,
-                    key_space: KeySpace::Unsecurified,
-                }
-                .matches(&p)
-            })
-            .map(|fi| fi.index)
-            .max()
-    }
-
-    fn max_account_mfa(&self, factor_source_id: FactorSourceIDFromHash) -> Option<HDPathComponent> {
-        self.securified_accounts_on_network
-            .clone()
-            .into_iter()
-            .flat_map(|e: SecurifiedAccount| {
-                e.highest_derivation_path_index(
-                    factor_source_id,
-                    AssertMatches {
-                        network_id: self.network_id,
-                        key_kind: CAP26KeyKind::TransactionSigning,
-                        entity_kind: CAP26EntityKind::Account,
-                        key_space: KeySpace::Securified,
-                    },
-                )
-            })
-            .max()
-    }
-
-    fn max_identity_mfa(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-    ) -> Option<HDPathComponent> {
-        self.securified_identities_on_network
-            .clone()
-            .into_iter()
-            .flat_map(|e: SecurifiedPersona| {
-                e.highest_derivation_path_index(
-                    factor_source_id,
-                    AssertMatches {
-                        network_id: self.network_id,
-                        key_kind: CAP26KeyKind::TransactionSigning,
-                        entity_kind: CAP26EntityKind::Identity,
-                        key_space: KeySpace::Securified,
-                    },
-                )
-            })
-            .max()
-    }
-
-    pub fn next(
-        &self,
-        agnostic_path: DerivationPreset,
-        factor_source_id: FactorSourceIDFromHash,
-    ) -> Result<Option<HDPathComponent>> {
-        let last = if agnostic_path == DerivationPreset::AccountVeci {
-            self.max_account_veci(factor_source_id)
-        } else if agnostic_path == DerivationPreset::AccountMfa {
-            self.max_account_mfa(factor_source_id)
-        } else if agnostic_path == DerivationPreset::IdentityMfa {
-            self.max_identity_mfa(factor_source_id)
-        } else if agnostic_path == DerivationPreset::IdentityVeci {
-            self.max_identity_veci(factor_source_id)
-        } else {
-            unreachable!("Unrecognized agnostic_path: {:?}", agnostic_path);
-        };
-        let Some(last) = last else { return Ok(None) };
-        last.add_one().map(Some)
-    }
-}
-
-#[derive(Debug)]
-pub struct NextDerivationEntityIndexWithLocalOffsetsForFactorSource {
-    #[allow(dead_code)]
-    factor_source_id: FactorSourceIDFromHash,
-    local_offsets: RwLock<HashMap<DerivationPreset, HDPathValue>>,
-}
-impl NextDerivationEntityIndexWithLocalOffsetsForFactorSource {
-    pub fn empty(factor_source_id: FactorSourceIDFromHash) -> Self {
-        Self {
-            factor_source_id,
-            local_offsets: RwLock::new(HashMap::new()),
-        }
-    }
-    pub fn reserve(&self, agnostic_path: DerivationPreset) -> HDPathValue {
-        let mut binding = self.local_offsets.write().unwrap();
-        if let Some(existing) = binding.get_mut(&agnostic_path) {
-            let free = *existing;
-            existing.add_assign(1);
-            free
-        } else {
-            binding.insert(agnostic_path, 1);
-            0
-        }
-    }
-}
-
-#[derive(Default, Debug)]
-pub struct NextDerivationEntityIndexWithLocalOffsets {
-    local_offsets_per_factor_source: RwLock<
-        HashMap<FactorSourceIDFromHash, NextDerivationEntityIndexWithLocalOffsetsForFactorSource>,
-    >,
-}
-
-impl NextDerivationEntityIndexWithLocalOffsets {
-    pub fn reserve(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-        agnostic_path: DerivationPreset,
-    ) -> HDPathValue {
-        let mut binding = self.local_offsets_per_factor_source.write().unwrap();
-        if let Some(for_factor) = binding.get_mut(&factor_source_id) {
-            for_factor.reserve(agnostic_path)
-        } else {
-            let new =
-                NextDerivationEntityIndexWithLocalOffsetsForFactorSource::empty(factor_source_id);
-            let next = new.reserve(agnostic_path);
-            binding.insert(factor_source_id, new);
-            next
-        }
-    }
-}
+use crate::prelude::*;
 
 pub struct NextDerivationEntityIndexAssigner {
     #[allow(dead_code)]
     network_id: NetworkID,
     profile_analyzing: NextDerivationEntityIndexProfileAnalyzingAssigner,
     local_offsets: NextDerivationEntityIndexWithLocalOffsets,
-}
-
-pub enum OffsetFromCache {
-    /// Finding max amongst already loaded (and removed) from cache, saved
-    /// locally
-    FindMaxInRemoved {
-        pf_found_in_cache: IndexMap<FactorSourceIDFromHash, FactorInstances>,
-    },
-    /// Known max by having peeked into the cache earlier.
-    KnownMax {
-        instance: HierarchicalDeterministicFactorInstance,
-    },
-}
-
-impl OffsetFromCache {
-    pub fn next(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-        index_agnostic_path: IndexAgnosticPath,
-    ) -> Result<Option<HDPathComponent>> {
-        let Some(m) = self._max(factor_source_id, index_agnostic_path) else {
-            return Ok(None);
-        };
-        m.add_one().map(Some)
-    }
-
-    fn _max(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-        index_agnostic_path: IndexAgnosticPath,
-    ) -> Option<HDPathComponent> {
-        match self {
-            Self::FindMaxInRemoved { pf_found_in_cache } => pf_found_in_cache
-                .get(&factor_source_id)
-                .cloned()
-                .unwrap_or_default()
-                .into_iter()
-                .filter(|f| f.agnostic_path() == index_agnostic_path)
-                .map(|f| f.derivation_path().index)
-                .max(),
-            Self::KnownMax { instance } => {
-                assert_eq!(instance.factor_source_id(), factor_source_id);
-                assert_eq!(instance.agnostic_path(), index_agnostic_path);
-                Some(instance.derivation_path().index)
-            }
-        }
-    }
 }
 
 impl NextDerivationEntityIndexAssigner {
