+++ src/factor_instances_provider/next_index_assigner/next_derivation_entity_index_with_local_offsets.rs
@@ -1,0 +1,27 @@
+use crate::prelude::*;
+
+#[derive(Default, Debug)]
+pub struct NextDerivationEntityIndexWithLocalOffsets {
+    local_offsets_per_factor_source: RwLock<
+        HashMap<FactorSourceIDFromHash, NextDerivationEntityIndexWithLocalOffsetsForFactorSource>,
+    >,
+}
+
+impl NextDerivationEntityIndexWithLocalOffsets {
+    pub fn reserve(
+        &self,
+        factor_source_id: FactorSourceIDFromHash,
+        agnostic_path: DerivationPreset,
+    ) -> HDPathValue {
+        let mut binding = self.local_offsets_per_factor_source.write().unwrap();
+        if let Some(for_factor) = binding.get_mut(&factor_source_id) {
+            for_factor.reserve(agnostic_path)
+        } else {
+            let new =
+                NextDerivationEntityIndexWithLocalOffsetsForFactorSource::empty(factor_source_id);
+            let next = new.reserve(agnostic_path);
+            binding.insert(factor_source_id, new);
+            next
+        }
+    }
+}
