+++ src/factor_instances_provider/next_index_assigner/next_derivation_entity_index_with_local_offsets_for_factor_source.rs
@@ -1,0 +1,30 @@
+use std::ops::AddAssign;
+
+use crate::prelude::*;
+
+#[derive(Debug)]
+pub struct NextDerivationEntityIndexWithLocalOffsetsForFactorSource {
+    #[allow(dead_code)]
+    factor_source_id: FactorSourceIDFromHash,
+    local_offsets: RwLock<HashMap<DerivationPreset, HDPathValue>>,
+}
+
+impl NextDerivationEntityIndexWithLocalOffsetsForFactorSource {
+    pub fn empty(factor_source_id: FactorSourceIDFromHash) -> Self {
+        Self {
+            factor_source_id,
+            local_offsets: RwLock::new(HashMap::new()),
+        }
+    }
+    pub fn reserve(&self, agnostic_path: DerivationPreset) -> HDPathValue {
+        let mut binding = self.local_offsets.write().unwrap();
+        if let Some(existing) = binding.get_mut(&agnostic_path) {
+            let free = *existing;
+            existing.add_assign(1);
+            free
+        } else {
+            binding.insert(agnostic_path, 1);
+            0
+        }
+    }
+}
